fixing hubSpot crm card issue (#220)
* fixing hubspot crm card issue

* chnage lead 360 column type

* Simplify logic

* Pass updatedAt for the case when lead is present in db (via cobalt) but analytics pipeline hasn't run and the stats haven't been calculated

---------

+++ src/main/java/com/sharefable/api/entity/Lead360.java
@@ -23,7 +23,7 @@ public class Lead360 extends EntityBase {
   private Integer timeSpentSec;
   private Timestamp lastInteractedAt;
   private Integer completionPercentage;
-  private Integer ctaClickRate;
+  private Double ctaClickRate;
 
   @Transient
   public static Lead360 Empty(Long tourId) {
@@ -34,7 +34,7 @@ public class Lead360 extends EntityBase {
       .timeSpentSec(0)
       .lastInteractedAt(null)
       .completionPercentage(0)
-      .ctaClickRate(0)
+      .ctaClickRate(0.0)
       .build();
   }
 
@@ -47,7 +47,7 @@ public class Lead360 extends EntityBase {
       .timeSpentSec(0)
       .lastInteractedAt(null)
       .completionPercentage(0)
-      .ctaClickRate(0)
+      .ctaClickRate(0.0)
       .build();
   }
 }
